pyproject.toml
[project]
name = "circuit-checker-v3"
version = "0.0.0"
description = "Prototype circuit checker"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "pytest>=8.3.5",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
